ui: extract shared sparkline component
Release note: None
